fix wrong type for factor, my bad

+++ headers/UI/UI_Element.h
@@ -35,7 +35,7 @@ void UI_Element_match_size(UI_Element *element, UI_Element *frame, float scale_x
 //Returns true if (x, y) is in element, false otherwise
 int UI_element_collision(UI_Element *element, int x, int y, float scale_x, float scale_y);
 //Positions the element at (x, y) based on the anchor and the scales
-void UI_Element_set_position(UI_Element *element, int x, int y, int x_position_factor, int y_position_factor, float scale_x, float scale_y, int anchor);
+void UI_Element_set_position(UI_Element *element, int x, int y, float x_position_factor, float y_position_factor, float scale_x, float scale_y, int anchor);
 //Draws the element inline
 void UI_Element_render_inline(UI_Element *element, float scale_x, float scale_y);
 //Draws the element outline
